pyproject.toml
[tool.poetry]
name = "fastapi-starter-kit"
version = "0.1.0"
description = ""
authors = ["Rafiqul Hasan <shopnilsazal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.3"
pre-commit = "^3.7.0"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
